Add new Product and ProductRepo classes
This update introduces two new classes, Product and ProductRepo, into the java project. The Product class represents individual products with a name and price. The ProductRepo class manages a list of products, and includes a method to add new products. Minor updates were made to the vcs.xml and misc.xml files within the .idea directory.