Update OrderRepo implementations and modify ShopService
The OrderRepo implementations (OrderListRepo and OrderMapRepo) have been updated to specify the type of order container in system messages for clearer logging. Moreover, ProductRepo and OrderRepo have been moved to instance variables in the ShopService class to improve code organization and readability. This also eliminates the need to instantiate these objects in the placeNewOrder method, making it more efficient.