Update methods to handle list of products
The methods in the Order and OrderRepo classes have been updated to handle lists of products instead of dealing with the ProductRepo class. This revision provides more clarity and better error handling in the ShopService class, particularly in the placeNewOrder method. Furthermore, print statements in Repo classes have been improved to denote the data structure used.